refactoring to pull out the common defintions into there own file.
